Result type changed

+++ Lesson1/HomeWork1/src/main/java/Program.java
@@ -7,7 +7,7 @@ public class Program {
     public static void main(String[] args) {
         List<Integer> numbers = new ArrayList<>(List.of(1, 2, 3, 4, 5, 6, 7, 8, 9, 10));
 
-        int result = (int) numbers.stream()
+        double result = numbers.stream()
                 .filter(n -> (n % 2) == 0)
                 .mapToInt(Integer::intValue)
                 .average().getAsDouble();
